requirements.txt
flask==3.1.2
flask-cors==6.0.1
langchain==0.1.9
langchain-core==0.1.53
langchain-community==0.0.38
langchain-huggingface==0.0.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
pymupdf==1.24.10
python-dotenv==1.0.1
tqdm==4.67.1
